backend/requirements.txt
bcrypt==5.0.0
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
filelock==3.20.1
Flask==3.1.2
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
nltk==3.9.2
numpy==2.2.6
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pycparser==2.23
PyJWT==2.10.1
pyotp==2.9.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
qrcode==8.2
regex==2025.11.3
requests==2.32.5
requests-file==3.0.1
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
SQLAlchemy==2.0.45
textblob==0.19.0
threadpoolctl==3.6.0
tldextract==5.3.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
Werkzeug==3.1.4
zxcvbn==4.5.0
